Make specific sdks unpublishable
I want to expose sdk-specific stuff to plugins in a different way that doesn't require publishing increasing amounts of nuget packages for specific toolchains. The Environment Settings system I added is a step towards that, which is why this commit has Sdk.Devkitpro set an environment setting from the DEVKITPRO environment variable.